# server/docs/swagger.yaml
definitions:
...
  robot.RobAdminLog:
    properties:
      content:
        description: 操作说明
        type: string
      id:
        description: id字段
        type: integer
      operPath:
        description: 操作pathname
        type: string
      operPost:
        description: 提交数据
        type: string
      operTime:
        description: 操作时间
        type: integer
      userIp:
        description: 操作ip
        type: string
      userName:
        description: 操作人
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.6.2
paths:
...
  /robAdminLog/createRobAdminLog:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建robAdminLog表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/robot.RobAdminLog'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建robAdminLog表
      tags:
      - RobAdminLog
  /robAdminLog/deleteRobAdminLog:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除robAdminLog表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/robot.RobAdminLog'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除robAdminLog表
      tags:
      - RobAdminLog
  /robAdminLog/deleteRobAdminLogByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除robAdminLog表
      tags:
      - RobAdminLog
  /robAdminLog/findRobAdminLog:
    get:
      consumes:
      - application/json
      parameters:
      - description: 操作说明
        in: query
        name: content
        type: string
      - description: id字段
        in: query
        name: id
        type: integer
      - description: 操作pathname
        in: query
        name: operPath
        type: string
      - description: 提交数据
        in: query
        name: operPost
        type: string
      - description: 操作时间
        in: query
        name: operTime
        type: integer
      - description: 操作ip
        in: query
        name: userIp
        type: string
      - description: 操作人
        in: query
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询robAdminLog表
      tags:
      - RobAdminLog
  /robAdminLog/getRobAdminLogList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取robAdminLog表列表
      tags:
      - RobAdminLog
  /robAdminLog/updateRobAdminLog:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新robAdminLog表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/robot.RobAdminLog'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新robAdminLog表
      tags:
      - RobAdminLog
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"